requirements.txt
Flask==2.0.1
keras==2.9.0
numpy==1.19.5
Pillow==9.2.0
opencv-python
tensorflow==2.5.0